requirements.txt
Flask
hdbcli
python-dotenv
requests
pandas
openpyxl